Introduce setting which controls whether strict rule validation should be applied
